# Various tweaks to the PercentEncoding API implementation