--- a/src/Dubzer.WhatwgUrl/PercentEncoding.cs
+++ b/src/Dubzer.WhatwgUrl/PercentEncoding.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Buffers;
 using System.Diagnostics;
-using System.Runtime.CompilerServices;
 using System.Text;
 using Dubzer.WhatwgUrl.BclInternal;
 
@@ -189,13 +188,14 @@ internal static class PercentEncoding
                     switch (notRequireEncodeUnifiedIndex)
                     {
                         case -1: // The entire string up to the end needs to be encoded
-                            foreach (ref readonly var c in processing) EncodeToUtf8HexWithPercent(c, ref vsb);
+                            foreach (ref readonly var c in processing)
+                                EncodeToUtf8HexWithPercent(c, ref vsb);
                             return true;
-
                         default: // The string partially requires character encoding
                             Debug.Assert(notRequireEncodeUnifiedIndex != 0);
 
-                            foreach (ref readonly var c in processing[..notRequireEncodeUnifiedIndex]) EncodeToUtf8HexWithPercent(c, ref vsb);
+                            foreach (ref readonly var c in processing[..notRequireEncodeUnifiedIndex])
+                                EncodeToUtf8HexWithPercent(c, ref vsb);
 
                             processing = processing[notRequireEncodeUnifiedIndex..];
                             continue;
@@ -212,13 +212,19 @@ internal static class PercentEncoding
 
         static int MakeUnifiedIndex(int firstSetIndex, int secondSetIndex)
         {
-            return firstSetIndex == -1 ? secondSetIndex : secondSetIndex == -1 ? firstSetIndex : firstSetIndex < secondSetIndex ? firstSetIndex : secondSetIndex;
+            if (firstSetIndex == -1)
+                return secondSetIndex;
+
+            if (secondSetIndex == -1)
+                return firstSetIndex;
+
+            return Math.Min(firstSetIndex, secondSetIndex);
         }
 
-        static int MakeInvertedUnifiedIndex(int firstSetIndex, int secondSetIndex)
-        {
-            return firstSetIndex == -1 || secondSetIndex == -1 ? -1 : firstSetIndex > secondSetIndex ? firstSetIndex : secondSetIndex;
-        }
+        static int MakeInvertedUnifiedIndex(int firstSetIndex, int secondSetIndex) =>
+            firstSetIndex == -1 || secondSetIndex == -1
+                ? -1
+                : Math.Max(firstSetIndex, secondSetIndex);
     }
 
     private static readonly SearchValues<char> SpecialHandlingChars = SearchValues.Create(['\\', '%']);
@@ -242,7 +248,7 @@ internal static class PercentEncoding
 
 #endif
 
-    internal static void AppendEncoded(Rune input, StringBuilder sb, in ReadOnlySpan<byte> c0Set)
+    internal static void AppendEncoded(Rune input, StringBuilder sb, ReadOnlySpan<byte> c0Set)
     {
         var c = input.ToChar();
         if ((uint) c > 0x7E || c0Set[(byte) c] != 0)
@@ -255,7 +261,7 @@ internal static class PercentEncoding
         }
     }
 
-    internal static void AppendEncoded(char input, StringBuilder sb, in ReadOnlySpan<byte> c0Set)
+    internal static void AppendEncoded(char input, StringBuilder sb, ReadOnlySpan<byte> c0Set)
     {
         if ((uint) input > 0x7E || ((byte) input < c0Set.Length && c0Set[(byte) input] != 0))
         {
@@ -317,7 +323,7 @@ internal static class PercentEncoding
     }
 
     // Inlined Rune.TryEncodeToUtf8
-    // Better to unify on one builder type
+    // TODO: Better to unify on one builder type
     private static void EncodeToUtf8HexWithPercent(char c, ref ValueStringBuilder vsb)
     {
         switch ((uint) c)
@@ -446,7 +452,7 @@ internal static class PercentEncoding
                         case 1: // Found the %
                             prevByte = ref b;
 
-                            if (!IsHexChar(b))
+                            if (!HexConverter.IsHexChar(b))
                             {
                                 output[outputIndex++] = (byte) '%'; // % recovery
 
@@ -454,12 +460,12 @@ internal static class PercentEncoding
                                 goto case 0;
                             }
 
-                            hexByte = (byte) (FromChar(b) << 4);
+                            hexByte = (byte) (HexConverter.FromChar(b) << 4);
                             processingState = 2;
                             continue;
 
                         case 2: // Found first hex character after %
-                            if (!IsHexChar(b))
+                            if (!HexConverter.IsHexChar(b))
                             {
                                 output[outputIndex + 0] = (byte) '%'; // % recovery
                                 output[outputIndex + 1] = prevByte; // previous byte recovery
@@ -469,7 +475,7 @@ internal static class PercentEncoding
                                 goto case 0;
                             }
 
-                            hexByte |= (byte) FromChar(b);
+                            hexByte |= (byte) HexConverter.FromChar(b);
                             output[outputIndex++] = hexByte;
                             processingState = 0;
                             continue;
@@ -502,65 +508,4 @@ internal static class PercentEncoding
             ArrayPool<byte>.Shared.Return(sourceBytes);
         }
     }
-
-    // TODO: Maybe you should move all below to Util
-    // Source: HexConverter.cs#L433
-
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private static bool IsHexChar(int c)
-    {
-        if (IntPtr.Size == 8)
-        {
-            // This code path, when used, has no branches and doesn't depend on cache hits,
-            // so it's faster and does not vary in speed depending on input data distribution.
-            // We only use this logic on 64-bit systems, as using 64 bit values would otherwise
-            // be much slower than just using the lookup table anyway (no hardware support).
-            // The magic constant 18428868213665201664 is a 64 bit value containing 1s at the
-            // indices corresponding to all the valid hex characters (ie. "0123456789ABCDEFabcdef")
-            // minus 48 (ie. '0'), and backwards (so from the most significant bit and downwards).
-            // The offset of 48 for each bit is necessary so that the entire range fits in 64 bits.
-            // First, we subtract '0' to the input digit (after casting to uint to account for any
-            // negative inputs). Note that even if this subtraction underflows, this happens before
-            // the result is zero-extended to ulong, meaning that `i` will always have upper 32 bits
-            // equal to 0. We then left shift the constant with this offset, and apply a bitmask that
-            // has the highest bit set (the sign bit) if and only if `c` is in the ['0', '0' + 64) range.
-            // Then we only need to check whether this final result is less than 0: this will only be
-            // the case if both `i` was in fact the index of a set bit in the magic constant, and also
-            // `c` was in the allowed range (this ensures that false positive bit shifts are ignored).
-            ulong i = (uint) c - '0';
-            var shift = 18428868213665201664UL << (int) i;
-            var mask = i - 64;
-
-            return (long) (shift & mask) < 0;
-        }
-
-        return FromChar(c) != 0xFF;
-    }
-
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private static int FromChar(int c)
-    {
-        return c >= CharToHexLookup.Length ? 0xFF : CharToHexLookup[c];
-    }
-
-    /// <summary>Map from an ASCII char to its hex value, e.g. arr['b'] == 11. 0xFF means it's not a hex digit.</summary>
-    private static ReadOnlySpan<byte> CharToHexLookup =>
-    [
-        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 15
-        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 31
-        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 47
-        0x0, 0x1, 0x2, 0x3, 0x4, 0x5, 0x6, 0x7, 0x8, 0x9, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 63
-        0xFF, 0xA, 0xB, 0xC, 0xD, 0xE, 0xF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 79
-        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 95
-        0xFF, 0xa, 0xb, 0xc, 0xd, 0xe, 0xf, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 111
-        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 127
-        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 143
-        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 159
-        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 175
-        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 191
-        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 207
-        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 223
-        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 239
-        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF // 255
-    ];
 }