… src/Dubzer.WhatwgUrl/BclInternal/HexConverter.cs