This PR has an alternate implementation for the `PercentEncoding` methods, the rest of the implementation is only affected in places where the changes were incompatible with the existing API. The implementation has been tested with .NET 8/9/10P2.

Perhaps some of the optimizations are unnecessary for now, but will be interesting to explore

It may be worth exploring other parts of the library for further improvement

---

```
BenchmarkDotNet v0.14.1-nightly.20250107.205, Windows 10 (10.0.19045.5608/22H2/2022Update)
12th Gen Intel Core i9-12900K 3.19GHz, 1 CPU, 24 logical and 16 physical cores
.NET SDK 10.0.100-preview.2.25164.34
  [Host]     : .NET 9.0.3 (9.0.325.11113), X64 RyuJIT AVX2
  Job-GWEAYY : .NET 10.0.0 (10.0.25.16302), X64 RyuJIT AVX2
  Job-TMCVYQ : .NET 8.0.13 (8.0.1325.6609), X64 RyuJIT AVX2
  Job-MTDGPX : .NET 9.0.3 (9.0.325.11113), X64 RyuJIT AVX2
```

---
Benchmark cases:

- `https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEgeGsYGEW9MWNDQ8Rv9vnsBcXgwqzjVCsUbBLDk9MGO30cXWmwsVmAMm0BHTLeSy1CJ9RghsDiDimj5tN0uDbb4M6Hoq7C8uBwBhRzn2yBV0UfkomGbHMqSUp2aqiwYzlmwBtEdR6O91Vqu2_yCw0E8LOpfmSyU0KE5QYfDPylF6obl2ps1ma83Z0Jt2A/s1270/E02737868%20Fiber%20Creative%20Templates%20Q2%202022_5%20Gig%20Valentine's%20Day%20Social%20Posts_1270x760_Apr22%20v02_Final%20(1).jpg`

- `https://www.anrdoezrs.net/links/8029122/type/dlg/sid/PEOAnneHathawayStruttedThroughParisintheHollywoodLovedFlatteringJeansTrendYouCanGetfor38ssavoyFasAff13927714202302I/https://www.madewell.com/the-perfect-vintage-wide-leg-jean-in-keller-wash-pocket-edition-NJ251.htmlhttps://www.madewell.com/the-perfect-vintage-wide-leg-jean-in-keller-wash-pocket-edition-NJ251.html`

---
**Before:**

| Runtime       | input                   |            Mean |    Ratio |       Gen0 |       Gen1 |  Allocated | Alloc Ratio |
| ------------- | ----------------------- | --------------: | -------: | ---------: | ---------: | ---------: | ----------: |
| .NET 10.0     | ?                       | 50,824,044.3 ns |     0.98 |  5500.0000 |          - | 86468990 B |        1.00 |
| .NET 8.0      | ?                       | 51,864,199.3 ns |     1.00 |  5500.0000 |          - | 86468986 B |        1.00 |
| .NET 9.0      | ?                       | 49,514,969.2 ns |     0.95 |  5500.0000 |          - | 86469448 B |        1.00 |
|               |                         |                 |          |            |            |            |             |
| **.NET 10.0** | **http(...).jpg [388]** |  **2,157.7 ns** | **0.94** | **0.1411** |      **-** | **2232 B** |    **1.00** |
| .NET 8.0      | http(...).jpg [388]     |      2,307.2 ns |     1.00 |     0.1411 |          - |     2232 B |        1.00 |
| .NET 9.0      | http(...).jpg [388]     |      2,096.6 ns |     0.91 |     0.1411 |          - |     2232 B |        1.00 |
|               |                         |                 |          |            |            |            |             |
| **.NET 10.0** | **http(...)html [367]** |    **410.7 ns** | **0.94** | **0.1225** | **0.0005** | **1928 B** |    **1.00** |
| .NET 8.0      | http(...)html [367]     |        436.8 ns |     1.00 |     0.1225 |     0.0005 |     1928 B |        1.00 |
| .NET 9.0      | http(...)html [367]     |        435.8 ns |     1.00 |     0.1225 |     0.0005 |     1928 B |        1.00 |




---

**After:**

| Runtime       | input                   |                Mean |    Ratio |          Gen0 |       Gen1 |       Allocated | Alloc Ratio |
| ------------- | ----------------------- | ------------------: | -------: | ------------: | ---------: | --------------: | ----------: |
| **.NET 10.0** | **?**                   | **38,429,853.8 ns** | **0.93** | **5461.5385** |      **-** | **84438.38 KB** |    **1.00** |
| .NET 8.0      | ?                       |     41,413,516.1 ns |     1.00 |     5500.0000 |          - |     84438.35 KB |        1.00 |
| .NET 9.0      | ?                       |     39,546,171.8 ns |     0.95 |     5461.5385 |          - |     84438.38 KB |        1.00 |
|               |                         |                     |          |               |            |                 |             |
| **.NET 10.0** | **http(...).jpg [388]** |      **1,540.3 ns** | **0.96** |    **0.1411** |      **-** |     **2.18 KB** |    **1.00** |
| .NET 8.0      | http(...).jpg [388]     |          1,598.6 ns |     1.00 |        0.1411 |          - |         2.18 KB |        1.00 |
| .NET 9.0      | http(...).jpg [388]     |          1,509.0 ns |     0.94 |        0.1411 |          - |         2.18 KB |        1.00 |
|               |                         |                     |          |               |            |                 |             |
| **.NET 10.0** | **http(...)html [367]** |        **335.5 ns** | **0.83** |    **0.1225** | **0.0005** |     **1.88 KB** |    **1.00** |
| .NET 8.0      | http(...)html [367]     |            404.9 ns |     1.00 |        0.1225 |     0.0005 |         1.88 KB |        1.00 |
| .NET 9.0      | http(...)html [367]     |            359.7 ns |     0.89 |        0.1225 |     0.0005 |         1.88 KB |        1.00 |